+++ finagle-example/src/main/scala/com/twitter/finagle/example/thrift/ThriftServicePerEndpointExample.scala
@@ -7,7 +7,7 @@ import com.twitter.finagle.service.{RetryExceptionsFilter, RetryPolicy, TimeoutF
 import com.twitter.finagle.util.DefaultTimer
 import com.twitter.util.{Await, Duration, Future, Throw, Try}
 
-object ThriftServiceIfaceExample {
+object ThriftServicePerEndpointExample {
   def main(args: Array[String]): Unit = {
     // See the docs at https://twitter.github.io/finagle/guide/Protocols.html#using-finagle-thrift
     //#thriftserverapi