+++ finagle-thrift/src/main/scala/com/twitter/finagle/thrift/service/ServicePerEndpointBuilder.scala
@@ -1,14 +1,14 @@
 package com.twitter.finagle.thrift.service
 
 import com.twitter.finagle.Service
-import com.twitter.finagle.thrift.{RichClientParam, ThriftClientRequest, ThriftServiceIface}
+import com.twitter.finagle.thrift.{RichClientParam, ThriftClientRequest}
 
 /**
  * Typeclass ServicePerEndpointBuilder[T] creates T-typed interfaces from thrift clients.
  * Scrooge generates implementations of this builder.
  */
 trait ServicePerEndpointBuilder[
-  ServicePerEndpoint <: ThriftServiceIface.Filterable[
+  ServicePerEndpoint <: Filterable[
     ServicePerEndpoint
   ]] {