+++ finagle-thrift/src/main/scala/com/twitter/finagle/thrift/service/ThriftServicePerEndpoint.scala
@@ -1,82 +1,13 @@
-package com.twitter.finagle.thrift
+package com.twitter.finagle.thrift.service
 
-import com.twitter.app.GlobalFlag
-import com.twitter.conversions.StorageUnitOps._
-import com.twitter.finagle._
+import com.twitter.finagle.{Filter, Service, SimpleFilter}
 import com.twitter.finagle.service.ResponseClassifier
-import com.twitter.finagle.stats.{NullStatsReceiver, StatsReceiver}
-import com.twitter.finagle.thrift.service.{
-  ThriftCodec,
-  ThriftMethodStatsHandler,
-  ThriftSourcedException
-}
-import com.twitter.scrooge._
-import com.twitter.util._
+import com.twitter.finagle.stats.StatsReceiver
+import com.twitter.finagle.thrift.{RichClientParam, ThriftClientRequest, ThriftMethodStats}
+import com.twitter.scrooge.ThriftMethod
+import com.twitter.util.Future
 import org.apache.thrift.TApplicationException
 import org.apache.thrift.protocol.TProtocolFactory
-
-object maxReusableBufferSize
-    extends GlobalFlag[StorageUnit](
-      16.kilobytes,
-      "Max size (bytes) for ThriftServiceIface reusable transport buffer"
-    )
-
-/**
- * Typeclass ServiceIfaceBuilder[T] creates T-typed interfaces from thrift clients.
- * Scrooge generates implementations of this builder.
- */
-@deprecated("Use com.twitter.finagle.thrift.service.ServicePerEndpointBuilder", "2017-11-13")
-trait ServiceIfaceBuilder[ServiceIface <: ThriftServiceIface.Filterable[ServiceIface]] {
-
-  /**
-   * Build a client ServiceIface wrapping a binary thrift service.
-   *
-   * @param thriftService An underlying thrift service that works on byte arrays.
-   * @param clientParam RichClientParam wraps client params [[com.twitter.finagle.thrift.RichClientParam]].
-   */
-  def newServiceIface(
-    thriftService: Service[ThriftClientRequest, Array[Byte]],
-    clientParam: RichClientParam
-  ): ServiceIface
-
-  @deprecated("Use com.twitter.finagle.thrift.RichClientParam", "2017-08-16")
-  def newServiceIface(
-    thriftService: Service[ThriftClientRequest, Array[Byte]],
-    pf: TProtocolFactory = Protocols.binaryFactory(),
-    stats: StatsReceiver = NullStatsReceiver,
-    responseClassifier: ResponseClassifier = ResponseClassifier.Default
-  ): ServiceIface = {
-    val clientParam =
-      RichClientParam(pf, clientStats = stats, responseClassifier = responseClassifier)
-    newServiceIface(thriftService, clientParam)
-  }
-
-  @deprecated("Use com.twitter.finagle.thrift.RichClientParam", "2017-08-16")
-  def newServiceIface(
-    thriftService: Service[ThriftClientRequest, Array[Byte]],
-    pf: TProtocolFactory,
-    stats: StatsReceiver
-  ): ServiceIface = {
-    val clientParam = RichClientParam(pf, clientStats = stats)
-    newServiceIface(thriftService, clientParam)
-  }
-}
-
-/**
- * A typeclass to construct a MethodIface by wrapping a ServiceIface.
- * This is a compatibility constructor to replace an existing Future interface
- * with one built from a ServiceIface.
- *
- * Scrooge generates implementations of this builder.
- */
-@deprecated("Use com.twitter.finagle.thrift.service.MethodPerEndpointBuilder", "2017-11-13")
-trait MethodIfaceBuilder[ServiceIface, MethodIface] {
-
-  /**
-   * Build a FutureIface wrapping a ServiceIface.
-   */
-  def newMethodIface(serviceIface: ServiceIface): MethodIface
-}
 
 /**
  * Construct Service interface for a Thrift method.
@@ -95,15 +26,15 @@ trait MethodIfaceBuilder[ServiceIface, MethodIface] {
  * }
  * }}}
  *
- * the `Service` interface, or `ServiceIface`, is
+ * the `Service` interface, or `ServicePerEndpoint`, is
  * {{{
- * trait LoggerServiceIface {
+ * trait LoggerServicePerEndpoint {
  *   val log: com.twitter.finagle.Service[Logger.Log.Args, Logger.Log.SuccessType]
  *   val getLogSize: com.twitter.finagle.Service[Logger.GetLogSize.Args, Logger.GetLogSize.SuccessType]
  * }
  * }}}
  *
- * and the method interface, or `MethodIface`, is
+ * and the method interface, or `MethodPerEndpoint`, is
  * {{{
  * trait Logger[Future] {
  *   def log(message: String, logLevel: Int): Future[String]
@@ -113,10 +44,7 @@ trait MethodIfaceBuilder[ServiceIface, MethodIface] {
  *
  * Service interfaces can be modified and composed with Finagle `Filters`.
  */
-object ThriftServiceIface { // TODO: Rename ThriftServicePerEndpoint and move to com.twitter.finagle.thrift.service
-
-  @deprecated("Use com.twitter.finagle.thrift.service.Filterable", "2017-11-06")
-  type Filterable[+T] = com.twitter.finagle.thrift.service.Filterable[T]
+object ThriftServicePerEndpoint {
 
   /**
    * Build a Service from a given Thrift method.
@@ -135,21 +63,6 @@ object ThriftServiceIface { // TODO: Rename ThriftServicePerEndpoint and move to
     stats
       .andThen(ThriftCodec.filter(method, clientParam.protocolFactory))
       .andThen(thriftService)
-  }
-
-  @deprecated("Use com.twitter.finagle.thrift.RichClientParam", "2017-08-16")
-  def apply(
-    method: ThriftMethod,
-    thriftService: Service[ThriftClientRequest, Array[Byte]],
-    pf: TProtocolFactory,
-    stats: StatsReceiver,
-    responseClassifier: ResponseClassifier
-  ): Service[method.Args, method.SuccessType] = {
-    apply(
-      method,
-      thriftService,
-      RichClientParam(pf, clientStats = stats, responseClassifier = responseClassifier)
-    )
   }
 
   def apply(